requirements.txt
Pillow==9.5.0
pytesseract==0.3.10
pyinstaller==5.10.1
clipboard==0.0.4